Remove Vaadin 8 and run tests with Flow using Saucelabs
Also removes support for old browsers: IE8-IE10, PhantomJS

Needs to uses Flow snapshot until 0.1.16 is available with a concurrency fix for https://github.com/vaadin/flow/issues/2677
