+++ doc/license.html
@@ -120,7 +120,6 @@
     </p>
     <ul>
         <li>Selenium 3.6.0 &ndash; Apache License 2.0</li>
-        <li>PhantomJSDriver 1.4.2 &ndash; The BSD 2-Clause License</li>
         <li>Javassist 3.18.1-GA &ndash; Apache License 2.0</li>
         <li>JUnit 4.12 &ndash; Common Public License Version 1.0</li>
         <li>org.json &ndash; The JSON License</li>
