+++ vaadin-testbench-integration-tests/src/main/java/com/vaadin/testUI/ElementQueryView.java
@@ -1,0 +1,16 @@
+package com.vaadin.testUI;
+
+import com.vaadin.router.Route;
+import com.vaadin.ui.html.Div;
+import com.vaadin.ui.html.NativeButton;
+
+@Route("ElementQueryView")
+public class ElementQueryView extends Div {
+
+    public ElementQueryView() {
+        for (int i = 0; i < 10; i++) {
+            add(new Div(new NativeButton("Button " + i)));
+        }
+    }
+
+}
