+++ vaadin-testbench-integration-tests/src/test/java/com/vaadin/tests/elements/ElementScreenCompareIT.java
@@ -2,23 +2,23 @@ package com.vaadin.tests.elements;
 
 import org.junit.Assert;
 import org.junit.Test;
+import org.openqa.selenium.By;
 
-import com.vaadin.testUI.ElementQueryUI;
-import com.vaadin.testbench.By;
+import com.vaadin.testUI.ElementQueryView;
 import com.vaadin.testbench.TestBenchElement;
+import com.vaadin.tests.MultiBrowserTest;
 
 public class ElementScreenCompareIT extends MultiBrowserTest {
 
     @Override
-    protected String getDeploymentPath() {
-        return "/" + ElementQueryUI.class.getSimpleName();
+    protected Class<ElementQueryView> getTestView() {
+        return ElementQueryView.class;
     }
 
     @Test
     public void elementCompareScreen() throws Exception {
         openTestURL();
-        TestBenchElement button4 = (TestBenchElement) findElements(
-                By.className("v-button")).get(4);
+        TestBenchElement button4 = $(NativeButtonElement.class).get(4);
 
         Assert.assertTrue(button4.compareScreen("button4"));
         TestBenchElement layout = (TestBenchElement) button4
