+++ vaadin-testbench-integration-tests/src/test/java/com/vaadin/tests/elements/ExecuteJavascriptIT.java
@@ -3,26 +3,24 @@ package com.vaadin.tests.elements;
 import org.junit.Assert;
 import org.junit.Test;
 
-import com.vaadin.testUI.ElementQueryUI;
-import com.vaadin.testbench.By;
+import com.vaadin.testUI.ElementQueryView;
 import com.vaadin.testbench.TestBenchElement;
+import com.vaadin.tests.MultiBrowserTest;
+import com.vaadin.ui.Component;
 
 public class ExecuteJavascriptIT extends MultiBrowserTest {
 
     @Override
-    protected String getDeploymentPath() {
-        return "/" + ElementQueryUI.class.getSimpleName();
+    protected Class<? extends Component> getTestView() {
+        return ElementQueryView.class;
     }
 
     @Test
     public void elementCompareScreen() throws Exception {
         openTestURL();
 
-        TestBenchElement button = (TestBenchElement) findElements(
-                By.className("v-button")).get(0);
-
-        Long offsetTop = (Long) executeScript("return arguments[0].offsetTop",
-                button);
+        TestBenchElement button = $(NativeButtonElement.class).first();
+        Long offsetTop = button.getPropertyDouble("offsetTop").longValue();
         Assert.assertEquals(Long.valueOf(0), offsetTop);
     }
 }
