+++ vaadin-testbench-integration-tests/src/test/java/com/vaadin/tests/elements/NativeButtonElement.java
@@ -1,0 +1,10 @@
+package com.vaadin.tests.elements;
+
+import com.vaadin.testbench.TestBenchElement;
+import com.vaadin.testbench.elementsbase.Element;
+
+@Element("button")
+public  class NativeButtonElement extends TestBenchElement {
+
+}
+
