+++ vaadin-testbench/src/main/java/com/vaadin/testbench/parallel/DefaultBrowserFactory.java
@@ -46,29 +46,9 @@ public class DefaultBrowserFactory implements TestBenchBrowserFactory {
             desiredCapabilities.setVersion(version);
             desiredCapabilities.setPlatform(platform);
             break;
-        case PHANTOMJS:
-            desiredCapabilities = DesiredCapabilities.phantomjs();
-            desiredCapabilities.setVersion(version);
-            desiredCapabilities.setPlatform(platform);
-            break;
         case SAFARI:
             desiredCapabilities = DesiredCapabilities.safari();
             desiredCapabilities.setVersion(version);
-            desiredCapabilities.setPlatform(platform);
-            break;
-        case IE8:
-            desiredCapabilities = DesiredCapabilities.internetExplorer();
-            desiredCapabilities.setVersion("8");
-            desiredCapabilities.setPlatform(platform);
-            break;
-        case IE9:
-            desiredCapabilities = DesiredCapabilities.internetExplorer();
-            desiredCapabilities.setVersion("9");
-            desiredCapabilities.setPlatform(platform);
-            break;
-        case IE10:
-            desiredCapabilities = DesiredCapabilities.internetExplorer();
-            desiredCapabilities.setVersion("10");
             desiredCapabilities.setPlatform(platform);
             break;
         case IE11:
