+++ vaadin-testbench/src/test/java/com/vaadin/testbench/ParametersThroughPropertiesTest.java
@@ -1,7 +1,6 @@
 package com.vaadin.testbench;
 
 import org.junit.Assert;
-import org.junit.Test;
 
 public class ParametersThroughPropertiesTest {
