+++ vaadin-testbench/src/test/java/com/vaadin/testbench/RetryRuleBeforeTest.java
@@ -5,8 +5,6 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
-import com.vaadin.testbench.RetryRule;
-
 
 public class RetryRuleBeforeTest {
     private int beforeExecutedCount = 0;
