+++ vaadin-testbench/src/test/java/com/vaadin/testbench/screenshot/ReferenceNameGeneratorTest.java
@@ -74,19 +74,6 @@ public class ReferenceNameGeneratorTest {
     }
 
     @Test
-    public void testGenerateName_barPhantomJSinCapabilities_returnsGeneratedName() {
-        Capabilities phantom = createNiceMock(Capabilities.class);
-        expect(phantom.getPlatform()).andReturn(Platform.MAC);
-        expect(phantom.getBrowserName()).andReturn("phantomjs");
-        expect(phantom.getVersion())
-                .andReturn("phantomjs-1.8.1+ghostdriver-1.0.2");
-        replay(phantom);
-        String name = rng.generateName("bar", phantom);
-        assertEquals("bar_mac_phantomjs_1", name);
-        verify(phantom);
-    }
-
-    @Test
     public void testGenerateName_shotEdgeinCapabilities_returnsGeneratedName() {
         Capabilities chrome = createNiceMock(Capabilities.class);
         expect(chrome.getPlatform()).andReturn(Platform.XP);
